--- core/src/main/java/io/temporal/samples/nexus_messaging/callerpattern/caller/CallerWorkflow.java
@@ -1,11 +1,0 @@
-package io.temporal.samples.nexus_messaging.callerpattern.caller;
-
-import io.temporal.workflow.WorkflowInterface;
-import io.temporal.workflow.WorkflowMethod;
-import java.util.List;
-
-@WorkflowInterface
-public interface CallerWorkflow {
-  @WorkflowMethod
-  List<String> run(String userId);
-}
