--- core/src/main/java/io/temporal/samples/nexus_messaging/callerpattern/service/Language.java
@@ -1,11 +1,0 @@
-package io.temporal.samples.nexus_messaging.callerpattern.service;
-
-public enum Language {
-  ARABIC,
-  CHINESE,
-  ENGLISH,
-  FRENCH,
-  HINDI,
-  PORTUGUESE,
-  SPANISH
-}
