--- core/src/main/java/io/temporal/samples/nexus_messaging/ondemandpattern/service/Language.java
@@ -1,11 +1,0 @@
-package io.temporal.samples.nexus_messaging.ondemandpattern.service;
-
-public enum Language {
-  ARABIC,
-  CHINESE,
-  ENGLISH,
-  FRENCH,
-  HINDI,
-  PORTUGUESE,
-  SPANISH
-}
