+++ core/src/main/java/io/temporal/samples/nexusmessaging/callerpattern/caller/CallerWorkflow.java
@@ -1,0 +1,11 @@
+package io.temporal.samples.nexusmessaging.callerpattern.caller;
+
+import io.temporal.workflow.WorkflowInterface;
+import io.temporal.workflow.WorkflowMethod;
+import java.util.List;
+
+@WorkflowInterface
+public interface CallerWorkflow {
+  @WorkflowMethod
+  List<String> run(String userId);
+}
