+++ core/src/main/java/io/temporal/samples/nexusmessaging/ondemandpattern/service/Language.java
@@ -1,0 +1,11 @@
+package io.temporal.samples.nexusmessaging.ondemandpattern.service;
+
+public enum Language {
+  ARABIC,
+  CHINESE,
+  ENGLISH,
+  FRENCH,
+  HINDI,
+  PORTUGUESE,
+  SPANISH
+}
